daemon: Storage trait
