+++ xelis_daemon/src/core/storage/difficulty_provider.rs
@@ -5,7 +5,7 @@ use xelis_common::{block::BlockHeader, crypto::hash::Hash};
 
 use crate::core::error::BlockchainError;
 
-use super::Storage;
+use super::{SledStorage, Storage};
 
 
 // this trait is useful for P2p to check itself the validty of a chain
@@ -20,25 +20,25 @@ pub trait DifficultyProvider {
 }
 
 #[async_trait]
-impl DifficultyProvider for Storage {
+impl DifficultyProvider for SledStorage {
     async fn get_height_for_block(&self, hash: &Hash) -> Result<u64, BlockchainError> {
         self.get_height_for_block_hash(hash).await
     }
 
     async fn get_timestamp_for_block(&self, hash: &Hash) -> Result<u128, BlockchainError> {
-        self.get_timestamp_for_block(hash).await
+        self.get_timestamp_for_block_hash(hash).await
     }
 
     async fn get_difficulty_for_block(&self, hash: &Hash) -> Result<u64, BlockchainError> {
-        self.get_difficulty_for_block(hash)
+        self.get_difficulty_for_block_hash(hash)
     }
 
     async fn get_cumulative_difficulty_for_block(&self, hash: &Hash) -> Result<u64, BlockchainError> {
-        self.get_cumulative_difficulty_for_block(hash).await
+        self.get_cumulative_difficulty_for_block_hash(hash).await
     }
 
     async fn get_past_blocks_of<'a>(&'a self, hash: &Hash) -> Result<Arc<Vec<Hash>>, BlockchainError> {
-        self.get_past_blocks_of(hash).await
+        self.get_past_blocks_for_block_hash(hash).await
     }
 
     async fn get_block_header<'a>(&'a self, hash: &Hash) -> Result<Arc<BlockHeader>, BlockchainError> {
